put load_model and load_data
